+++ src/Equinor.Procosys.Preservation.Infrastructure/PreservationContext.cs
@@ -1,7 +1,5 @@
 ﻿using System.Linq;
 using System.Reflection;
-using System.Threading;
-using System.Threading.Tasks;
 using Equinor.Procosys.Preservation.Domain;
 using Equinor.Procosys.Preservation.Domain.AggregateModels.JourneyAggregate;
 using Equinor.Procosys.Preservation.Domain.AggregateModels.ModeAggregate;
@@ -9,31 +7,20 @@ using Equinor.Procosys.Preservation.Domain.AggregateModels.PersonAggregate;
 using Equinor.Procosys.Preservation.Domain.AggregateModels.ProjectAggregate;
 using Equinor.Procosys.Preservation.Domain.AggregateModels.RequirementTypeAggregate;
 using Equinor.Procosys.Preservation.Domain.AggregateModels.ResponsibleAggregate;
-using Equinor.Procosys.Preservation.Domain.Events;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.ChangeTracking;
 
 namespace Equinor.Procosys.Preservation.Infrastructure
 {
-    public class PreservationContext : DbContext, IUnitOfWork, IReadOnlyContext
+    public class PreservationContext : DbContext, IReadOnlyContext
     {
-        private readonly IEventDispatcher _eventDispatcher;
         private readonly IPlantProvider _plantProvider;
-        private readonly ITimeService _timeService;
-        private readonly ICurrentUserProvider _currentUserProvider;
 
         public PreservationContext(
             DbContextOptions<PreservationContext> options,
-            IEventDispatcher eventDispatcher,
-            IPlantProvider plantProvider,
-            ITimeService timeService,
-            ICurrentUserProvider currentUserProvider)
+            IPlantProvider plantProvider)
             : base(options)
         {
-            _eventDispatcher = eventDispatcher;
             _plantProvider = plantProvider;
-            _timeService = timeService;
-            _currentUserProvider = currentUserProvider;
         }
 
         protected override void OnModelCreating(ModelBuilder modelBuilder)
@@ -41,19 +28,6 @@ namespace Equinor.Procosys.Preservation.Infrastructure
             base.OnModelCreating(modelBuilder);
             modelBuilder.ApplyConfigurationsFromAssembly(Assembly.GetExecutingAssembly());
             SetGlobalPlantFilter(modelBuilder);
-        }
-
-        private void SetGlobalPlantFilter(ModelBuilder modelBuilder)
-        {
-            // Set global query filter on entities inheriting from SchemaEntityBase
-            // https://gunnarpeipman.com/ef-core-global-query-filters/
-            foreach (var type in TypeProvider.GetEntityTypes(typeof(IDomainMarker).GetTypeInfo().Assembly, typeof(SchemaEntityBase)))
-            {
-                typeof(PreservationContext)
-                .GetMethod(nameof(PreservationContext.SetGlobalQueryFilter))
-                ?.MakeGenericMethod(type)
-                .Invoke(this, new object[] { modelBuilder });
-            }
         }
 
         public static DateTimeKindConverter DateTimeKindConverter { get; } = new DateTimeKindConverter();
@@ -75,20 +49,17 @@ namespace Equinor.Procosys.Preservation.Infrastructure
         public virtual DbSet<FieldValue> FieldValues { get; set; }
         public virtual DbSet<Action> Actions { get; set; }
 
-        public override async Task<int> SaveChangesAsync(CancellationToken cancellationToken = default)
+        private void SetGlobalPlantFilter(ModelBuilder modelBuilder)
         {
-            await DispatchEvents(cancellationToken);
-            await SetAuditData();
-            return await base.SaveChangesAsync(cancellationToken);
-        }
-
-        private async Task DispatchEvents(CancellationToken cancellationToken = default)
-        {
-            var entities = ChangeTracker
-                .Entries<EntityBase>()
-                .Where(x => x.Entity.DomainEvents != null && x.Entity.DomainEvents.Any())
-                .Select(x => x.Entity);
-            await _eventDispatcher.DispatchAsync(entities, cancellationToken);
+            // Set global query filter on entities inheriting from SchemaEntityBase
+            // https://gunnarpeipman.com/ef-core-global-query-filters/
+            foreach (var type in TypeProvider.GetEntityTypes(typeof(IDomainMarker).GetTypeInfo().Assembly, typeof(SchemaEntityBase)))
+            {
+                typeof(PreservationContext)
+                .GetMethod(nameof(PreservationContext.SetGlobalQueryFilter))
+                ?.MakeGenericMethod(type)
+                .Invoke(this, new object[] { modelBuilder });
+            }
         }
 
         public void SetGlobalQueryFilter<T>(ModelBuilder builder) where T : SchemaEntityBase =>
@@ -97,28 +68,5 @@ namespace Equinor.Procosys.Preservation.Infrastructure
             .HasQueryFilter(e => e.Schema == _plantProvider.Plant);
 
         public IQueryable<TEntity> QuerySet<TEntity>() where TEntity : class => Set<TEntity>().AsNoTracking();
-
-        private async Task SetAuditData()
-        {
-            var auditables = ChangeTracker.Entries<IAuditable>().Where(x => x.State == EntityState.Added || x.State == EntityState.Modified);
-            if (auditables.Any())
-            {
-                var now = _timeService.GetCurrentTimeUtc();
-                var currentUser = await _currentUserProvider.GetCurrentUserAsync();
-
-                foreach (var entry in auditables)
-                {
-                    switch (entry.State)
-                    {
-                        case EntityState.Added:
-                            entry.Entity.SetCreated(now, currentUser.Id);
-                            break;
-                        case EntityState.Modified:
-                            entry.Entity.SetModified(now, currentUser.Id);
-                            break;
-                    }
-                }
-            }
-        }
     }
 }
