+++ src/Equinor.Procosys.Preservation.WebApi/Seeding/Seeder.cs
@@ -15,6 +15,7 @@ namespace Equinor.Procosys.Preservation.WebApi.Seeding
 {
     public class Seeder : IHostedService
     {
+        private static readonly Person s_seederUser = new Person(new Guid("12345678-1234-1234-1234-123456789123"), "Angus", "MacGyver");
         private readonly IServiceScopeFactory _serviceProvider;
 
         public Seeder(IServiceScopeFactory serviceProvider) => _serviceProvider = serviceProvider;
@@ -27,12 +28,26 @@ namespace Equinor.Procosys.Preservation.WebApi.Seeding
 
                 using (var dbContext = new PreservationContext(
                     scope.ServiceProvider.GetRequiredService<DbContextOptions<PreservationContext>>(),
-                    scope.ServiceProvider.GetRequiredService<IEventDispatcher>(),
-                    plantProvider,
-                    scope.ServiceProvider.GetRequiredService<ITimeService>(),
-                    scope.ServiceProvider.GetRequiredService<ICurrentUserProvider>()))
+                    plantProvider))
                 {
-                    var unitOfWork = dbContext;
+                    // If the seeder user exists in the database, it's already been seeded. Don't seed again.
+                    if (await dbContext.Persons.AnyAsync(p => p.Oid == s_seederUser.Oid))
+                    {
+                        return;
+                    }
+
+                    /* 
+                     * Add the initial seeder user. Don't do this through the UnitOfWork as this expects/requires the current user to exist in the database.
+                     * This is the first user that is added to the database and will not get "Created" and "CreatedBy" data.
+                     */
+                    dbContext.Persons.Add(s_seederUser);
+                    await dbContext.SaveChangesAsync(cancellationToken);
+
+                    var unitOfWork = new UnitOfWork(
+                        dbContext,
+                        scope.ServiceProvider.GetRequiredService<IEventDispatcher>(),
+                        scope.ServiceProvider.GetRequiredService<ITimeService>(),
+                        new SeederUserProvider());
                     var personRepository = new PersonRepository(dbContext);
                     var journeyRepository = new JourneyRepository(dbContext);
                     var modeRepository = new ModeRepository(dbContext);
@@ -63,7 +78,7 @@ namespace Equinor.Procosys.Preservation.WebApi.Seeding
 
                     var steps = (await journeyRepository.GetAllAsync()).SelectMany(x => x.Steps).ToList();
                     var requirementDefinitions = (await requirementTypeRepository.GetAllAsync()).SelectMany(x => x.RequirementDefinitions).ToList();
-                    
+
                     projects.AddTags(100, plantProvider.Plant, steps, requirementDefinitions);
                     await unitOfWork.SaveChangesAsync(cancellationToken);
                 }
@@ -74,7 +89,7 @@ namespace Equinor.Procosys.Preservation.WebApi.Seeding
 
         private class SeederUserProvider : ICurrentUserProvider
         {
-            public Task<Person> GetCurrentUserAsync() => Task.FromResult(new Person(Guid.NewGuid(), "Seeder", "Seederson"));
+            public Task<Person> GetCurrentUserAsync() => Task.FromResult(s_seederUser);
         }
     }
 }
