+++ src/Equinor.Procosys.Preservation.WebApi/appsettings.Development.json
@@ -7,5 +7,5 @@
     }
   },
   "MigrateDatabase": true,
-  "SeedDummyData": false
+  "SeedDummyData": true
 }
